requirements.txt
plotly==4.2.0
jupytext==1.1.1
jupyter
notebook
pandas==0.23.0
statsmodels==0.9.0
scipy==1.1.0
patsy==0.5.1
numpy==1.16.0
plotly-geo
python-igraph
geopandas
pyshp
shapely
psutil
requests
networkx
squarify
scikit-image
sphinx
recommonmark
python-frontmatter
pathlib